ezclap lets go ive done something
A duck walked up to a lemonade stand
And he said to the man, running the stand
"Hey! (Bum bum bum) Got any grapes?"
The man said
"No we just sell lemonade. But it's cold
And it's fresh
And it's all home-made. Can I get you
Glass?"
The duck said,
"I'll pass".